% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfn.R
\name{DTLZ1}
\alias{DTLZ1}
\title{DTLZ1}
\usage{
DTLZ1(s)
}
\arguments{
\item{s}{a numeric vector of length 7 (all parameters must be bounded by \link{0,1})}
}
\value{
a numeric vector of length 3
}
\description{
One of the test functions from the \url{https://github.com/amirmasoudabdol/amosa}{AMOSA} package.
It is a 7 dimensional continuous optimization problem with 3 objectives and can be found in the paper:
\itemize{
\item Deb, Kalyanmoy, et al. "Scalable multi-objective optimization test problems." Proceedings of the 2002 Congress on Evolutionary Computation. CEC'02 (Cat. No. 02TH8600). Vol. 1. IEEE, 2002.
}
}
