% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfn.R
\name{DTLZ1_fromC}
\alias{DTLZ1_fromC}
\title{Call DTLZ1 from C}
\usage{
DTLZ1_fromC(s)
}
\arguments{
\item{s}{a numeric vector of length 7 (all parameters must be bounded by \link{0,1})}
}
\value{
a numeric vector of length 3
}
\description{
See \code{\link[amosa]{DTLZ1}} for details.
}
